Move BlockEncodingManager to presto-common
