+++ presto-hive/src/test/java/com/facebook/presto/hive/util/TestSerDeUtils.java
@@ -13,9 +13,9 @@
  */
 package com.facebook.presto.hive.util;
 
-import com.facebook.presto.block.BlockEncodingManager;
 import com.facebook.presto.common.block.Block;
 import com.facebook.presto.common.block.BlockBuilder;
+import com.facebook.presto.common.block.BlockEncodingManager;
 import com.facebook.presto.common.block.BlockEncodingSerde;
 import com.facebook.presto.common.block.BlockSerdeUtil;
 import com.facebook.presto.common.type.ArrayType;
