+++ presto-main/src/test/java/com/facebook/presto/block/AbstractTestBlock.java
@@ -17,6 +17,7 @@ import com.facebook.presto.common.block.AbstractMapBlock.HashTables;
 import com.facebook.presto.common.block.Block;
 import com.facebook.presto.common.block.BlockBuilder;
 import com.facebook.presto.common.block.BlockBuilderStatus;
+import com.facebook.presto.common.block.BlockEncodingManager;
 import com.facebook.presto.common.block.BlockEncodingSerde;
 import com.facebook.presto.common.block.DictionaryBlock;
 import com.facebook.presto.common.block.DictionaryId;
