+++ presto-main/src/test/java/com/facebook/presto/execution/TestMemoryRevokingScheduler.java
@@ -17,7 +17,7 @@ package com.facebook.presto.execution;
 import com.facebook.airlift.stats.CounterStat;
 import com.facebook.airlift.stats.TestingGcMonitor;
 import com.facebook.presto.Session;
-import com.facebook.presto.block.BlockEncodingManager;
+import com.facebook.presto.common.block.BlockEncodingManager;
 import com.facebook.presto.execution.TestSqlTaskManager.MockExchangeClientSupplier;
 import com.facebook.presto.execution.buffer.OutputBuffers;
 import com.facebook.presto.execution.executor.TaskExecutor;
