+++ presto-main/src/test/java/com/facebook/presto/execution/buffer/TestingPagesSerdeFactory.java
@@ -13,8 +13,8 @@
  */
 package com.facebook.presto.execution.buffer;
 
-import com.facebook.presto.block.BlockEncodingManager;
 import com.facebook.presto.common.Page;
+import com.facebook.presto.common.block.BlockEncodingManager;
 import com.facebook.presto.common.block.BlockEncodingSerde;
 import com.facebook.presto.spi.page.PageCompressor;
 import com.facebook.presto.spi.page.PageDecompressor;
