+++ presto-main/src/test/java/com/facebook/presto/spiller/TestTempStorageSingleStreamSpiller.java
@@ -13,9 +13,9 @@
  */
 package com.facebook.presto.spiller;
 
-import com.facebook.presto.block.BlockEncodingManager;
 import com.facebook.presto.common.Page;
 import com.facebook.presto.common.block.BlockBuilder;
+import com.facebook.presto.common.block.BlockEncodingManager;
 import com.facebook.presto.common.io.OutputStreamDataSink;
 import com.facebook.presto.common.type.Type;
 import com.facebook.presto.memory.context.LocalMemoryContext;
