+++ presto-main/src/test/java/com/facebook/presto/sql/TestExpressionInterpreter.java
@@ -13,8 +13,8 @@
  */
 package com.facebook.presto.sql;
 
-import com.facebook.presto.block.BlockEncodingManager;
 import com.facebook.presto.common.block.Block;
+import com.facebook.presto.common.block.BlockEncodingManager;
 import com.facebook.presto.common.block.BlockEncodingSerde;
 import com.facebook.presto.common.block.BlockSerdeUtil;
 import com.facebook.presto.common.type.ArrayType;
